requirements.txt
Django==4.2.7
psycopg2-binary==2.9.5
python-dotenv==0.21.0
gunicorn==20.1.0
Pillow==10.1.0
django-cors-headers==4.1.0
django-storages==1.14.2
boto3==1.33.12
djangorestframework==3.14.0
Markdown==3.5.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
whitenoise==6.2.0
django-admin-interface==0.28.1
django-apptemplates==1.5
